+++ lib/zoi/types/default.ex
@@ -24,6 +24,11 @@ defmodule Zoi.Types.Default do
   end
 
   defimpl Zoi.TypeSpec do
+    def spec(%Zoi.Types.Default{inner: schema, value: nil}, opts) do
+      inner_spec = Zoi.TypeSpec.spec(schema, opts)
+      quote(do: nil | unquote(inner_spec))
+    end
+
     def spec(%Zoi.Types.Default{inner: schema}, opts) do
       Zoi.TypeSpec.spec(schema, opts)
     end