+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.17.2 - 2026-03-15
+
+### Changed
+
+- `Zoi.type_spec/1` for `Zoi.struct/3` now resolves optional fields as nilable
+- `Zoi.type_spec/1` for `Zoi.default/3` with `nil` value now resolves as nilable
+
 ## 0.17.1 - 2026-02-19
 
 ### Changed